+++ engine-core/src/Heightmap.cpp
@@ -6,7 +6,7 @@ namespace Core
 
     Heightmap::Heightmap(const std::string& filepath)
     {
-        int width, height, nrChannels;
+        int32 width, height, nrChannels;
         stbi_set_flip_vertically_on_load(true);
         unsigned char* localBuffer = stbi_load(filepath.c_str(), &width, &height, &nrChannels, 0);
         stbi_set_flip_vertically_on_load(false);
@@ -24,16 +24,16 @@ namespace Core
             else
             {
                 //Save image into buffer
-                int count = 0;
-                for(int i = 0; i < width; i++)
+                uint32 count = 0;
+                for(int32 i = 0; i < width; i++)
                 {
-                    for(int j = 0; j < height; j++)
+                    for(int32 j = 0; j < height; j++)
                     {
                         float x = (float)localBuffer[count] / 255.0f;
                         float y = (float)localBuffer[count + 1] / 255.0f;
                         float z = (float)localBuffer[count + 2] / 255.0f;
                         _heightArray[i][j] = x + y + z;
-                        count+=3;
+                        count += 3;
                     }
                 }
                 LOG(INFO) << "Successfully load heightmap texture at: " << filepath << " (X: " << width << " | Y: " << height << " | Channel: " << nrChannels << ")";
@@ -47,7 +47,7 @@ namespace Core
         stbi_image_free(localBuffer);
     }
 
-    float Heightmap::GetValueAt(const unsigned int x, const unsigned int z) const
+    float Heightmap::GetValueAt(const uint16 x, const uint16 z) const
     {
         if(x < PLANE_SIZE && z < PLANE_SIZE)
             return _heightArray[x][z];
